% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_cdn.R
\name{reactCDN}
\alias{reactCDN}
\title{Dependencies for React}
\usage{
reactCDN(version = 17)
}
\arguments{
\item{version}{Version of React to use, if `NULL` uses the latest}
}
\description{
Includes React dependencies in a shiny application.
}
\keyword{internal}
